switched int to long, which leads to better performance in some cases

+++ test/temp/ArrayAccessorConst.hpp
@@ -25,7 +25,7 @@
 #include "common/DataTypes.hpp"
 #else
 #include <assert.h>
-using  integer_t     = int;
+using  integer_t     = long;
 #endif
